dtype int for z

+++ bgmm/distributions.py
@@ -88,7 +88,7 @@ class Multinomial(object):
 
     def sample(self, n=1):
         # Returns component assignments using evidence
-        z = np.zeros(self.pi.shape[0])
+        z = np.zeros(self.pi.shape[0], dtype='int64')
         for i, j in enumerate(self.pi):
             z[i] = np.where(np.random.multinomial(n=n, pvals=j))[0][0]
         return z
